More work on howwehelp carousel page.
Also added local cache so that user will see same 'page' state when refreshing browser. However, I need to make sure that on componentDidMount the app component sets Index as default state. Want users to always be on first page on a fresh load of the app. Figure this out somehow?